requirements.txt
feedparser>=6.0.11
beautifulsoup4>=4.12.3
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
groq>=0.9.0
streamlit>=1.35.0
